Move gpu_util.go to e2e/framework/gpu
