+++ test/e2e/framework/BUILD
@@ -14,7 +14,6 @@ go_library(
         "framework.go",
         "get-kubemark-resource-usage.go",
         "google_compute.go",
-        "gpu_util.go",
         "jobs_util.go",
         "kubelet_stats.go",
         "log_size_monitoring.go",
@@ -156,6 +155,7 @@ filegroup(
         ":package-srcs",
         "//test/e2e/framework/config:all-srcs",
         "//test/e2e/framework/ginkgowrapper:all-srcs",
+        "//test/e2e/framework/gpu:all-srcs",
         "//test/e2e/framework/ingress:all-srcs",
         "//test/e2e/framework/metrics:all-srcs",
         "//test/e2e/framework/podlogs:all-srcs",
